install mr

+++ bootstrap.sh
@@ -6,8 +6,18 @@ cp vimrc ~/.vimrc
 curl https://raw.githubusercontent.com/Shougo/neobundle.vim/master/bin/install.sh | sh
 
 # Setup emacs
-mv ~/.emacs.d ~/.emacs.d.bak
+if ! [ -e ~/.emacs.d ]
+then
+    mv ~/.emacs.d ~/.emacs.d.bak
 
-git clone --recursive http://github.com/syl20bnr/spacemacs ~/.emacs.d
-ln -s ~/.dotfiles/.spacemacs ~/.spacemacs
-ln -s ~/.dotfiles/spacemacs-private/* ~/.emacs.d/private/
+    git clone --recursive http://github.com/syl20bnr/spacemacs ~/.emacs.d
+    ln -s ~/.dotfiles/.spacemacs ~/.spacemacs
+    ln -s ~/.dotfiles/spacemacs-private/* ~/.emacs.d/private/
+fi
+
+# Setup mr
+pushd ~/usr/bin
+curl https://raw.githubusercontent.com/joeyh/myrepos/master/mr > mr
+chmod 755 mr
+
+popd
